Fix issue #22
When the same value is called for instance in an infinite loop the fade function resets, i.e. by calling `stop_fade()` and setting `percent_done` to `0`. Therefore, in my fix I first check if the color hasn't changed or the fader is still fading to the requested value. Otherwise, 'reset' and begin fading.